ekzamen/requirements.txt
attrs==17.4.0
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
coverage==4.4.2
dj-database-url==0.4.2
Django==1.11.8
django-ckeditor==5.4.0
django-heroku==0.2.0
django-js-asset==1.0.0
django-rest-framework==0.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
flake8==3.5.0
flake8-polyfill==1.0.2
gunicorn==19.7.1
idna==2.6
mando==0.6.4
mccabe==0.6.1
Pillow==5.0.0
pluggy==0.6.0
psycopg2==2.7.3.2
py==1.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.5.3
pytest==3.3.2
pytest-cov==2.5.1
pytest-django==3.1.2
pytz==2017.3
PyYAML==3.12
radon==2.2.0
requests==2.18.4
six==1.11.0
urllib3==1.22
whitenoise==3.3.1
xenon==0.5.4
